Fix Sonar issues
